Remove SingletonConnect.java, turn Connector.java to singleton
