+++ nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=/Users/macbook/Library/Application Support/NetBeans/18/build.properties
+user.properties.file=C:\\Users\\ADMIN\\AppData\\Roaming\\NetBeans\\19\\build.properties
