+++ nbproject/private/private.xml
@@ -3,9 +3,8 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/Users/macbook/Desktop/Code/mywork/CinemaBooking/src/LoginPage/Login.java</file>
-            <file>file:/Users/macbook/Desktop/Code/mywork/CinemaBooking/src/Dashboard/MainPage.java</file>
-            <file>file:/Users/macbook/Desktop/Code/mywork/CinemaBooking/src/LoginPage/SignUp.java</file>
+            <file>file:/C:/code/CinemaBooking/CinemaBooking/src/Dashboard/MainPage.java</file>
+            <file>file:/C:/code/CinemaBooking/CinemaBooking/src/Connector/Connector.java</file>
         </group>
     </open-files>
 </project-private>
