+++ src/Dashboard/MainPage.java
@@ -4,7 +4,6 @@
  */
 package Dashboard;
 import Connector.Connector;
-import Connector.SingletonConnect;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.ResultSet;
@@ -17,7 +16,7 @@ import java.sql.Statement;
  * @author ADMIN
  */
 public class MainPage extends javax.swing.JFrame {
-
+    
     /**
      * Creates new form MainPage
      */
@@ -266,7 +265,7 @@ public class MainPage extends javax.swing.JFrame {
     private void dashboardBtnActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_dashboardBtnActionPerformed
         // TODO add your handling code here:
         jPanel2.setVisible(true);
-        try (Connection conn = SingletonConnect.getInstance().getConnection(); Statement stmt = conn.createStatement()) {
+        try (Connection conn = Connector.getInstance().getConnection(); Statement stmt = conn.createStatement()) {
             String SQL ="SELECT DISTINCT M.movieName, C.cinemaName, M.genre, M.duration, M.directorName, S.scheduleDate, S.totalTicket, S.ticketsSold " +
                         "FROM Movies M " +
                         "JOIN Schedule S ON S.movieName = M.movieName " +
